+++ tests/testthat/test_oper_source.R
@@ -1,4 +1,3 @@
-library("photobiology")
 library("lubridate")
 
 context("source_spct")